+++ admin/views/transactions/tmpl/default_body.php
@@ -21,12 +21,12 @@ defined('_JEXEC') or die;
         'class' => 'width-100px'
     );
     ?>
-    <tr class="row<?php echo $i % 2; ?> <?php echo JHtml::_('crowdfundingbackend.transactionColor', $item->txn_status); ?>">
+    <tr class="row<?php echo $i % 2; ?> <?php echo JHtml::_('crowdfundingbackend.transactionColor', $item->txn_status); ?>" id="txn-row-<?php echo $item->id; ?>">
         <td class="center hidden-phone">
             <?php echo JHtml::_('grid.id', $i, $item->id); ?>
         </td>
         <td>
-            <a href="<?php echo JRoute::_('index.php?option=com_crowdfundingfinance&view=transaction&layout=edit&id=' . $item->id); ?>"><?php echo $item->txn_id; ?></a>
+            <a href="<?php echo JRoute::_('index.php?option=com_crowdfunding&view=transaction&layout=edit&id=' . $item->id); ?>"><?php echo $item->txn_id; ?></a>
             <?php if (!empty($item->parent_txn_id)) { ?>
                 <div class="small">
                     <?php echo $this->escape($item->parent_txn_id); ?>
@@ -40,7 +40,7 @@ defined('_JEXEC') or die;
                 <?php echo JHtmlString::truncate(strip_tags($item->project), 53); ?>
             </a>
         </td>
-        <td><?php echo JHtml::_('crowdfundingbackend.transactionAmount', $item, $this->amount, $this->currencies); ?></td>
+        <td><?php echo JHtml::_('crowdfundingbackend.transactionAmount', $item, $this->money, $this->currencies); ?></td>
         <td class="hidden-phone"><?php echo $item->txn_date; ?></td>
         <td class="hidden-phone">
             <?php echo $item->service_provider; ?>
@@ -49,7 +49,7 @@ defined('_JEXEC') or die;
             </div>
         </td>
         <td class="hidden-phone">
-            <?php echo $item->txn_status; ?>
+            <?php echo JHtml::_('select.genericlist', $this->paymentStatuses, 'txn_status', array('class' => 'js-txn-status', 'data-id' => $item->id), 'value', 'text', $item->txn_status); ?>
             <?php echo JHtml::_('crowdfundingbackend.reason', $item->status_reason); ?>
         </td>
         <td class="center hidden-phone">