+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Crowdfunding Finance Changelog
 ===============================
 
+###v2.4
+* Improved amount formatting.
+* It was made compatible with Crowdfunding v2.5.
+
 ###v2.3.1
 * Fixed an issue with Stripe in the plugin Crowdfunding - Payout Options.
 * Added Update server. Now, it will be possible to upgrade the extension via Joomla! Update Manager.